Item class and updated config
There is now an Item class to represent the items of the spreadsheet.
This class is basically a name and a item count.
Config has also been updated to be able to specify the columns in which
the items are described, as well as a starting row to ignore the titles
rows.